+++ modules/base.nix
@@ -77,6 +77,12 @@ in {
 
     services.nix-daemon.enable = true;
 
+    # WORKAROUND: Using MagicDNS (through nix-darwin) without setting a fallback
+    # DNS server leads to taking a lot longer to connect to the internet.
+    networking.dns = [ "1.1.1.1" ];
+
+    services.tailscale.magicDNS.enable = true;
+
     # WORKAROUND: `systemsetup -f -setremotelogin on` requires `Full Disk Access`
     # permission for the Application calling it
     system.activationScripts.extraActivation.text = ''